requirements.txt
pandas 
gdown
dvc
mlflow==2.2.2
notebook
numpy
matplotlib
seaborn
nltk
scikit-learn==1.2.2
ensure==1.0.2
python-box==6.0.2
pyYAML
tqdm
joblib
streamlit
python-dotenv
-e .